web-scanner/requirements.txt
requests
beautifulsoup4
colorama
urllib3
dnspython
python-nmap
sslyze
jinja2
tqdm
pyjwt
cryptography
websocket-client
graphql-core
pymongo
lxml
pyyaml
selenium
webdriver-manager
